Separate VS Code install and config/reest

+++ install-docker.sh
@@ -8,6 +8,12 @@ echo "***************************************************"
 
 set -uxo pipefail
 
+if command -v docker &> /dev/null
+then
+    echo "docker already installed, skipping"
+    exit 0
+fi
+
 # Install Docker [https://docs.docker.com/engine/install/ubuntu/]
 sudo apt purge docker.io docker-doc docker-compose docker-compose-v2 podman-docker containerd runc
